# src/gittest/swagger.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
servers: [
    {
      "url": "http://localhost:5000"
    }
  ]
paths:
  /api/v2/users:
    get:
      tags:
        - users
      summary: Get all users
      security:
        - jwt: []
      responses:
        '200':
          description: List of found users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: No users found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
  /api/v2/users/login:
    post:
      tags:
        - users
      summary: Login user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '404':
          description: Unsuccessful login
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
  /api/v2/users/register:
    post:
      tags:
        - users
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fio:
                  type: string
                email:
                  type: string
                password:
                  type: string
                money:
                  type: number
                address:
                  type: string
                phone:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User created
        '404':
          description: User not created
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
  /api/v2/users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the user
    get:
      tags:
        - users
      summary: Get user by ID
      security:
        - jwt: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
    delete:
      tags:
        - users
      summary: Delete user by ID
      security:
        - jwt: []
      responses:
        '200':
          description: User deleted
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
    patch:
      tags:
        - users
      summary: Update user by ID
      security:
        - jwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fio:
                  type: string
                email:
                  type: string
                password:
                  type: string
                money:
                  type: number
                address:
                  type: string
                phone:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User updated successfully
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
  /api/v2/users/{email}:
    get:
      tags:
        - users
      summary: Get user by email
      security:
        - jwt: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
          description: User's email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
...
  /api/v2/payments/users/{user_id}:
    get:
      tags:
        - payments
      summary: Get payments by user_id
      security:
        - jwt: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of found payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '401':
          description: Unauthorized
        '404':
          description: Payments not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    post:
      tags:
        - payments
      summary: Create payment by user_id
      security:
        - jwt: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '401':
          description: Unauthorized
        '404':
          description: Payments not created
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        payment_id:
          type: integer
        all_price:
          type: number
        state:
          type: string
        user_id:
          type: integer
      required:
        - payment_id
        - all_price
        - state
        - user_id
...
    User:
      type: object
      properties:
        user_id:
          type: integer
        fio:
          type: string
        email:
          type: string
        password:
          type: string
        money:
          type: number
        address:
          type: string
        phone:
          type: string
        role:
          type: string
          default: user
      required:
        - user_id
        - fio
        - email
        - password
        - money
        - address
        - phone 
        - role 
...
    Token:
      type: object
      properties:
        access_token:
          type: string
      required:
        - access_token